stock.volume( int to long) + loading stocks in list

+++ src/main/java/telran/java47/fintech/model/Stock.java
@@ -29,7 +29,7 @@ public class Stock {
 	Double lowV;
 	Double closeV;
 	Double adjCloseV;
-	Integer volume;
+	Long volume;
 	Boolean workDayOrNot;
 
 }
